refactor(quiz): decompõe QuizAnhangaLanding em telas + dados (#968)
QuizAnhangaLanding.tsx cruzou 1110 linhas misturando 4 responsabilidades
num só módulo. Decompõe sem mudança de comportamento:

- Dados (perguntas, perfis, labels de resumo) → data/quiz.ts
- Telas e subcomponentes → components/landings/quiz/ (named exports,
  imports relativos, no padrão de components/landings/orlando/)
- A página mantém hook de URL, tipo Stage, StageContent, reducer e o
  componente orquestrador (1110 → 279 linhas)

tests/quiz-questions.test.ts lê o texto-fonte; aponta para os novos
arquivos (data/quiz.ts e ResultScreen.tsx) sem perder cobertura.
