+++ components/landings/quiz/WhatsAppUpgrade.tsx
@@ -1,0 +1,87 @@
+import { useState } from 'react';
+import { getWhatsAppLink } from '../../../utils/whatsapp';
+
+interface WhatsAppUpgradeProps {
+    profileName: string;
+    mainDestName: string;
+    firstName: string;
+    baseWaUrl: string;
+}
+
+function maskPhone(v: string): string {
+    const d = v.replace(/\D/g, '').slice(0, 11);
+    if (d.length <= 2) return d;
+    if (d.length <= 6) return `(${d.slice(0, 2)}) ${d.slice(2)}`;
+    if (d.length <= 10) return `(${d.slice(0, 2)}) ${d.slice(2, 6)}-${d.slice(6)}`;
+    return `(${d.slice(0, 2)}) ${d.slice(2, 7)}-${d.slice(7)}`;
+}
+
+export function WhatsAppUpgrade({ profileName, mainDestName, firstName, baseWaUrl }: WhatsAppUpgradeProps) {
+    const [phone, setPhone] = useState('');
+    const [submitted, setSubmitted] = useState(false);
+
+    const digits = phone.replace(/\D/g, '');
+    const isValid = digits.length >= 10;
+
+    const waUrl = isValid && submitted
+        ? getWhatsAppLink(
+            `Oi! Sou ${firstName}. Descobri no Quiz da Anhangá que meu perfil é ${profileName} e meu próximo destino pode ser ${mainDestName}. Bora planejar essa viagem?`,
+            { appendTrackingRef: true },
+          )
+        : baseWaUrl;
+
+    if (submitted) {
+        return (
+            <div className="quiz-wa-upgrade quiz-wa-upgrade--done">
+                <p className="quiz-wa-upgrade-done">
+                    Perfeito! Agora é só abrir o WhatsApp.
+                </p>
+                <a
+                    className="btn-whatsapp btn-specialist quiz-btn quiz-btn-primary quiz-btn-lg"
+                    href={waUrl}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    data-tracking="result-quiz-whatsapp"
+                >
+                    Montar minha viagem
+                    <span className="quiz-arrow">→</span>
+                </a>
+            </div>
+        );
+    }
+
+    return (
+        <div className="quiz-wa-upgrade">
+            <p className="quiz-wa-upgrade-label">
+                Quer guardar seu perfil no WhatsApp?
+            </p>
+            <div className="quiz-wa-upgrade-row">
+                <input
+                    type="tel"
+                    className="quiz-wa-input"
+                    placeholder="(11) 99999-9999"
+                    value={phone}
+                    onChange={(e) => setPhone(maskPhone(e.target.value))}
+                    aria-label="Número de WhatsApp (opcional)"
+                />
+                <button
+                    type="button"
+                    className={`quiz-btn quiz-btn-primary ${isValid ? 'quiz-btn--ready' : ''}`}
+                    onClick={() => isValid && setSubmitted(true)}
+                    disabled={!isValid}
+                >
+                    Enviar
+                </button>
+            </div>
+            <a
+                className="btn-whatsapp btn-specialist quiz-wa-skip"
+                href={baseWaUrl}
+                target="_blank"
+                rel="noopener noreferrer"
+                data-tracking="result-quiz-skip"
+            >
+                Ir pro WhatsApp sem deixar número →
+            </a>
+        </div>
+    );
+}
